# chuckswapi/src/chuckswapi_api/Controllers/FeedControllerApi.cs
/*
 * Chuck Norris IO
 *
 * chucknorris.io is a free JSON API for hand curated Chuck Norris facts.  Chuck Norris facts are satirical factoids about martial artist and actor Chuck Norris that have become an Internet phenomenon and as a result have become widespread in popular culture. The 'facts' are normally absurd hyperbolic claims about Norris' toughness, attitude, virility, sophistication, and masculinity.  Chuck Norris facts have spread around the world, leading not only to translated versions, but also spawning localized versions mentioning country-specific advertisements and other Internet phenomena. Allusions are also sometimes made to his use of roundhouse kicks to perform seemingly any task, his large amount of body hair with specific regard to his beard, and his role in the action television series Walker, Texas Ranger.
 *
 * OpenAPI spec version: 2.0.0
 * Contact: m@matchilling.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using chuckswapi_api.Attributes;

using Microsoft.AspNetCore.Authorization;
using chuckswapi_api.Models;

namespace chuckswapi_api.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class FeedControllerApiController : ControllerBase
    { 
        /// <summary>
        /// dailyChuckJson
        /// </summary>
        
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        [HttpGet]
        [Route("//feed/daily-chuck")]
        [ValidateModelState]
        [SwaggerOperation("DailyChuckJsonUsingGET")]
        [SwaggerResponse(statusCode: 200, type: typeof(DailyChuck), description: "OK")]
        public virtual IActionResult DailyChuckJsonUsingGET()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(DailyChuck));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            string exampleJson = null;
            exampleJson = "{\n  \"issue_number\" : 0,\n  \"issues\" : [ {\n    \"date\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"joke_id\" : \"joke_id\"\n  }, {\n    \"date\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"joke_id\" : \"joke_id\"\n  } ]\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<DailyChuck>(exampleJson)
            : default(DailyChuck);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// dailyChuckJson
        /// </summary>
        
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        [HttpGet]
        [Route("//feed/daily-chuck.json")]
        [ValidateModelState]
        [SwaggerOperation("DailyChuckJsonUsingGET1")]
        [SwaggerResponse(statusCode: 200, type: typeof(DailyChuck), description: "OK")]
        public virtual IActionResult DailyChuckJsonUsingGET1()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(DailyChuck));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            string exampleJson = null;
            exampleJson = "{\n  \"issue_number\" : 0,\n  \"issues\" : [ {\n    \"date\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"joke_id\" : \"joke_id\"\n  }, {\n    \"date\" : \"2000-01-23T04:56:07.000+00:00\",\n    \"joke_id\" : \"joke_id\"\n  } ]\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<DailyChuck>(exampleJson)
            : default(DailyChuck);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// dailyChuckRss
        /// </summary>
        
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        [HttpGet]
        [Route("//feed/daily-chuck.xml")]
        [ValidateModelState]
        [SwaggerOperation("DailyChuckRssUsingGET1")]
        [SwaggerResponse(statusCode: 200, type: typeof(View), description: "OK")]
        public virtual IActionResult DailyChuckRssUsingGET1()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(View));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            string exampleJson = null;
            exampleJson = "";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<View>(exampleJson)
            : default(View);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// dailyChuckStats
        /// </summary>
        
        /// <response code="200">OK</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        [HttpGet]
        [Route("//feed/daily-chuck/stats")]
        [ValidateModelState]
        [SwaggerOperation("DailyChuckStatsUsingGET")]
        [SwaggerResponse(statusCode: 200, type: typeof(MailingListStatistic), description: "OK")]
        public virtual IActionResult DailyChuckStatsUsingGET()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(MailingListStatistic));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            string exampleJson = null;
            exampleJson = "{\n  \"cleaned_count\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"unsubscribe_count_since_send\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"member_count_since_send\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"click_rate\" : { },\n  \"cleaned_count_since_send\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"campaign_count\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"avg_unsub_rate\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"last_sub_date\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"target_sub_rate\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"unsubscribe_count\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"avg_sub_rate\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"campaign_last_sent\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"member_count\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"merge_field_count\" : {\n    \"and_increment\" : 6,\n    \"and_decrement\" : 0\n  },\n  \"last_unsub_date\" : \"2000-01-23T04:56:07.000+00:00\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<MailingListStatistic>(exampleJson)
            : default(MailingListStatistic);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}