# chuckswapi/src/chuckswapi_api/Models/AtomicInteger.cs
/*
 * Chuck Norris IO
 *
 * chucknorris.io is a free JSON API for hand curated Chuck Norris facts.  Chuck Norris facts are satirical factoids about martial artist and actor Chuck Norris that have become an Internet phenomenon and as a result have become widespread in popular culture. The 'facts' are normally absurd hyperbolic claims about Norris' toughness, attitude, virility, sophistication, and masculinity.  Chuck Norris facts have spread around the world, leading not only to translated versions, but also spawning localized versions mentioning country-specific advertisements and other Internet phenomena. Allusions are also sometimes made to his use of roundhouse kicks to perform seemingly any task, his large amount of body hair with specific regard to his beard, and his role in the action television series Walker, Texas Ranger.
 *
 * OpenAPI spec version: 2.0.0
 * Contact: m@matchilling.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace chuckswapi_api.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class AtomicInteger : IEquatable<AtomicInteger>
    { 
        /// <summary>
        /// Gets or Sets AndDecrement
        /// </summary>
        [DataMember(Name="and_decrement")]
        public int? AndDecrement { get; set; }

        /// <summary>
        /// Gets or Sets AndIncrement
        /// </summary>
        [DataMember(Name="and_increment")]
        public int? AndIncrement { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class AtomicInteger {\n");
            sb.Append("  AndDecrement: ").Append(AndDecrement).Append("\n");
            sb.Append("  AndIncrement: ").Append(AndIncrement).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((AtomicInteger)obj);
        }

        /// <summary>
        /// Returns true if AtomicInteger instances are equal
        /// </summary>
        /// <param name="other">Instance of AtomicInteger to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(AtomicInteger other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    AndDecrement == other.AndDecrement ||
                    AndDecrement != null &&
                    AndDecrement.Equals(other.AndDecrement)
                ) && 
                (
                    AndIncrement == other.AndIncrement ||
                    AndIncrement != null &&
                    AndIncrement.Equals(other.AndIncrement)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (AndDecrement != null)
                    hashCode = hashCode * 59 + AndDecrement.GetHashCode();
                    if (AndIncrement != null)
                    hashCode = hashCode * 59 + AndIncrement.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(AtomicInteger left, AtomicInteger right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(AtomicInteger left, AtomicInteger right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}