# chuckswapi/src/chuckswapi_api/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "chucknorris.io is a free JSON API for hand curated Chuck Norris facts.\n\nChuck Norris facts are satirical factoids about martial artist and actor Chuck Norris that have become an Internet phenomenon and as a result have become widespread in popular culture. The 'facts' are normally absurd hyperbolic claims about Norris' toughness, attitude, virility, sophistication, and masculinity.\n\nChuck Norris facts have spread around the world, leading not only to translated versions, but also spawning localized versions mentioning country-specific advertisements and other Internet phenomena. Allusions are also sometimes made to his use of roundhouse kicks to perform seemingly any task, his large amount of body hair with specific regard to his beard, and his role in the action television series Walker, Texas Ranger.",
    "version" : "2.0.0",
    "title" : "Chuck Norris IO",
    "termsOfService" : "https://api.chucknorris.io/",
    "contact" : {
      "name" : "Mathias Schilling",
      "url" : "https://www.matchilling.com",
      "email" : "m@matchilling.com"
    },
    "license" : {
      "name" : "GNU General Public License v3.0",
      "url" : "https://github.com/chucknorris-io/chuck-api/blob/master/LICENSE"
    }
  },
  "host" : "api.chucknorris.io",
  "basePath" : "/",
...
  "paths" : {
    "/feed/daily-chuck" : {
      "get" : {
        "tags" : [ "feed-controller" ],
        "summary" : "dailyChuckJson",
        "operationId" : "dailyChuckJsonUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/DailyChuck"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/feed/daily-chuck.json" : {
      "get" : {
        "tags" : [ "feed-controller" ],
        "summary" : "dailyChuckJson",
        "operationId" : "dailyChuckJsonUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/DailyChuck"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/feed/daily-chuck.xml" : {
      "get" : {
        "tags" : [ "feed-controller" ],
        "summary" : "dailyChuckRss",
        "operationId" : "dailyChuckRssUsingGET_1",
        "produces" : [ "text/xml", "application/rss+xml" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/View"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/feed/daily-chuck/stats" : {
      "get" : {
        "tags" : [ "feed-controller" ],
        "summary" : "dailyChuckStats",
        "operationId" : "dailyChuckStatsUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/MailingListStatistic"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
...
  },
  "definitions" : {
    "AtomicDouble" : {
      "type" : "object",
      "title" : "AtomicDouble"
    },
    "AtomicInteger" : {
      "type" : "object",
      "properties" : {
        "and_decrement" : {
          "type" : "integer",
          "format" : "int32"
        },
        "and_increment" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "title" : "AtomicInteger",
      "example" : {
        "and_increment" : 6,
        "and_decrement" : 0
      }
    },
    "DailyChuck" : {
      "type" : "object",
      "properties" : {
        "issue_number" : {
          "type" : "integer",
          "format" : "int64"
        },
        "issues" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DailyChuckIssue"
          }
        }
      },
      "title" : "DailyChuck",
      "example" : {
        "issue_number" : 0,
        "issues" : [ {
          "date" : "2000-01-23T04:56:07.000+00:00",
          "joke_id" : "joke_id"
        }, {
          "date" : "2000-01-23T04:56:07.000+00:00",
          "joke_id" : "joke_id"
        } ]
      }
    },
    "DailyChuckIssue" : {
      "type" : "object",
      "properties" : {
        "date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "joke_id" : {
          "type" : "string"
        }
      },
      "title" : "DailyChuckIssue",
      "example" : {
        "date" : "2000-01-23T04:56:07.000+00:00",
        "joke_id" : "joke_id"
      }
    },
...
    "MailingListStatistic" : {
      "type" : "object",
      "properties" : {
        "avg_sub_rate" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "avg_unsub_rate" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "campaign_count" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "campaign_last_sent" : {
          "type" : "string",
          "format" : "date-time"
        },
        "cleaned_count" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "cleaned_count_since_send" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "click_rate" : {
          "$ref" : "#/definitions/AtomicDouble"
        },
        "last_sub_date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "last_unsub_date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "member_count" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "member_count_since_send" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "merge_field_count" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "open_rate" : {
          "$ref" : "#/definitions/AtomicDouble"
        },
        "target_sub_rate" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "unsubscribe_count" : {
          "$ref" : "#/definitions/AtomicInteger"
        },
        "unsubscribe_count_since_send" : {
          "$ref" : "#/definitions/AtomicInteger"
        }
      },
      "title" : "MailingListStatistic",
      "example" : {
        "cleaned_count" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "unsubscribe_count_since_send" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "member_count_since_send" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "click_rate" : { },
        "cleaned_count_since_send" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "campaign_count" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "avg_unsub_rate" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "last_sub_date" : "2000-01-23T04:56:07.000+00:00",
        "target_sub_rate" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "unsubscribe_count" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "avg_sub_rate" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "campaign_last_sent" : "2000-01-23T04:56:07.000+00:00",
        "member_count" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "merge_field_count" : {
          "and_increment" : 6,
          "and_decrement" : 0
        },
        "last_unsub_date" : "2000-01-23T04:56:07.000+00:00"
      }
    },
...
    "View" : {
      "type" : "object",
      "properties" : {
        "content_type" : {
          "type" : "string"
        }
      },
      "title" : "View"
    }
  }
}